% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_type.R
\name{doc_type}
\alias{doc_type}
\title{Rmarkdown document output type}
\usage{
doc_type()
}
\value{
Output document type as a string: "pdf", "html", or "other".
}
\description{
Return the rmarkdown document output type as string.
}
\details{
For use in Rmarkdown documents. Returns the output document type as a string.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
