% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDBpub-package.R
\docType{package}
\name{FDBpub}
\alias{FDBpub}
\alias{FDBpub-package}
\title{Helpers for report writing}
\description{
The FDBpub package holds a few functions designed to
help with building and formatting reports in an Rmarkdown
workflow. It includes helpers for ggplot graphics, kable tables,
and maybe some other stuff.
}
\author{
\strong{Maintainer}: David Braze \email{davebraze@gmail.com}

}
\keyword{internal}
