% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Convenience wrapper for knitr::kable() et al}
\usage{
format_table(
  content,
  out_type,
  digits = 2,
  caption = " ",
  booktabs = TRUE,
  longtable = FALSE,
  group_labels = NULL,
  group_starts = NULL,
  group_ends = NULL,
  bootstrap_options = c("striped", "condensed"),
  latex_options = c("basic", "HOLD_position"),
  full_width = FALSE,
  font_size = 10
)
}
\arguments{
\item{content}{A data.frame or similar holding table content. It will be passed to kable().}

\item{out_type}{The output type of the table. Specially handled values are "html", "pdf", which yield tables in html or latex/pdf format, respectively. Any other value will yield a 'pandoc' table. See 'format' argument to kable().}

\item{digits}{Number of digits to right of decimal. See same name argument to kable().}

\item{caption}{The table caption. See same name argument to kable().}

\item{booktabs}{Enable latex package 'booktabs'. See same name argument to kable().}

\item{longtable}{Enable latex package 'longtable'. See same name argument to kable().}

\item{group_labels}{Labels assigned to groups of table rows. See kableExtra::pack_rows().}

\item{group_starts}{Start of each group of table rows. See kableExtra::pack_rows().}

\item{group_ends}{Start of each group of table rows. See kableExtra::pack_rows().}

\item{bootstrap_options}{Used for html tables. See kableExtra::kable_styling().}

\item{latex_options}{Used for pdf tables. See kableExtra::kable_styling().}

\item{full_width}{Should table be full page width? See kableExtra::kable_styling()}

\item{font_size}{Font size. See kableExtra::kable_styling()}
}
\value{
A formated table, per knitr::kable().
}
\description{
Convenience wrapper for knitr::kable() designed to handle styling contingent upon output type (pdf vs. html). In addition to kable(), this wraps several other functions: kableExtra::pack_rows(), kableExtra::kable_styling(), kableExtra::column_spec().
}
\details{
Convenience wrapper for knitr::kable() designed to handle styling contingent upon output type (pdf vs. html). Allows separate formatting options depending on output document type.
}
\seealso{
kable(), kable_styling(), pack_rows(), column_spec().
}
\author{
David Braze \email{davebraze@gmail.com}
}
