% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_draft.R
\name{is_draft}
\alias{is_draft}
\title{Is this a draft document?}
\usage{
is_draft()
}
\value{
Logical value: TRUE or FALSE.
}
\description{
Return value of YAML directive "draft:"
}
\details{
For use in Rmarkdown documents. Return value of YAML directive "draft:", either \code{true} or \code{false}.

Will return FALSE if "draft:" directive does not exist in rmarkdown file.

Will return FALSE if value of "draft:" directive is not a logical value (true or false).
}
\examples{

## May be used to conditionally watermark ggplots
library(ggplot2)
data(iris)
tmp <- ggplot(iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
  geom_point()

if (is_draft()) {watermark(tmp)} else {tmp}

## If called from within YAML block use FDBpub::is_draft().
FDBpub::is_draft()

}
\seealso{
\code{\link[rmarkdown]{metadata}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
