% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_fdb.R
\name{palette_fdb}
\alias{palette_fdb}
\title{A palette}
\usage{
palette_fdb()
}
\value{
A color palette as a vector.
}
\description{
A palette, or maybe a set of them.
}
\details{
Someday I will design my own color palette. Look
\href{https://blog.datawrapper.de/colors-for-data-vis-style-guides/}{here}
for some ideas about how to do that.

But, I think I may just adapt some work from the MetBrewer package.
}
\examples{

}
\seealso{
\code{\link[MetBrewer]{met.brewer}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
