% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fdbplot.R
\name{theme_fdbplot}
\alias{theme_fdbplot}
\title{FDB's stock ggplot2 theme}
\usage{
theme_fdbplot(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  rel_small = 10/12,
  rel_large = 15/12
)
}
\arguments{
\item{base_size}{Base font size in points. Defaults to 12.}

\item{base_family}{Base font family. No default.}

\item{base_line_size}{Base line width in points. Defaults to base_size/24.}

\item{base_rect_size}{Base line width for rectangles. Defaults to base_size/24.}

\item{rel_small}{Small font size. Defaults to 10/12.}

\item{rel_large}{Large font size. Defaults to 15/12.}
}
\description{
GGplot2 theme based on theme_gray().
}
\details{
GGplot2 theme based on theme_gray().

Use ggplot2::theme_update() to tweak this to specific client needs.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
