% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Watermark ggplot with "DRAFT"}
\usage{
watermark(
  gg,
  label = "DRAFT",
  size = 108,
  color = "black",
  alpha = 0.1,
  angle = 45,
  x = 0.5,
  y = 0.5,
  ...
)
}
\arguments{
\item{gg}{A ggplot object to be watermarked.}

\item{label}{Text to use as watermark. Defaults to "DRAFT".}

\item{size}{Text size in points. Defaults to 108.}

\item{color}{Watermark color. Defaults to "black".}

\item{alpha}{Watermark transparency. Defaults to 0.1.}

\item{angle}{Watermark angle. Defaults to 45°.}

\item{x}{x position of label. Defaults to 0.5.}

\item{y}{y position of label. Defaults to 0.5.}

\item{...}{Arguments passed to cowplot::draw_label()}
}
\value{
A ggplot2 object.
}
\description{
Watermark ggplot with "DRAFT" or other text.
}
\details{
Watermark ggplot with "DRAFT" or other text.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
