% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Watermark ggplot with "DRAFT"}
\usage{
watermark(
  plot,
  label = "DRAFT",
  size = 108,
  color = "black",
  alpha = 0.1,
  angle = 45,
  x = 0.5,
  y = 0.5,
  do = TRUE,
  ...
)
}
\arguments{
\item{plot}{A ggplot object to be watermarked.}

\item{label}{Text to use as watermark. Defaults to "DRAFT".}

\item{size}{Text size in points. Defaults to 108.}

\item{color}{Watermark color. Defaults to "black".}

\item{alpha}{Watermark transparency. Defaults to 0.1.}

\item{angle}{Watermark angle. Defaults to 45°.}

\item{x}{x position of label. Defaults to 0.5.}

\item{y}{y position of label. Defaults to 0.5.}

\item{do}{logical specifying whether to apply the watermark. Defaults to TRUE.}

\item{...}{Arguments passed to cowplot::draw_label()}
}
\value{
A ggplot2 object.
}
\description{
Watermark ggplot with "DRAFT" or other text.
}
\details{
Watermark ggplot with "DRAFT" or other text.
}
\examples{
library(ggplot2)
data(iris)
tmp <- ggplot(iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
  geom_point()
  watermark(tmp)

## alternately, in an rmarkdown file maybe do

watermark(tmp, do=is_draft())

}
\seealso{
\code{\link[cowplot]{draw_label}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
